backend/requirements.txt
aniso8601==1.3.0
asn1crypto==0.23.0
bcrypt==3.1.4
cffi==1.11.2
click==6.7
cryptography==2.1.3
Fabric3==1.13.1.post1
Flask==0.12.2
Flask-RESTful==0.3.6
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
Werkzeug==0.12.2
